[PATCH] paride cleanup and fixes (25/25)
paride has a ->private field for use by low-level drivers.  It's not
64bit-clean - it's int, but is used to store a pointer (bpck6).
Fix: replace with unsigned long, update typecasts in users.

+++ drivers/block/paride/paride.h
@@ -45,7 +45,7 @@ struct pi_adapter  {
 	int	saved_r0;	     /* saved port state */
 	int	saved_r2;	     /* saved port state */
 	int	reserved;	     /* number of ports reserved */
-	int	private;	     /* for protocol module */
+	unsigned long	private;     /* for protocol module */
 
 	wait_queue_head_t parq;     /* semaphore for parport sharing */
 	void	*pardev;	     /* pointer to pardevice */
